requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pyaudio==0.2.14
sounddevice==0.4.6
scipy==1.11.1
audioread==3.0.1

# Speech-to-Text
openai-whisper==20231117
ffmpeg-python==0.2.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Streamlit UI
streamlit==1.32.0
streamlit-audiorecorder==0.0.8

# Utilities
joblib==1.3.2
tqdm==4.65.0
Pillow==10.0.0